requirements.txt
pillow
python-dotenv
requests
langchain-google-genai
langchain-openai